Reestruturação pastas, responsivo,  mudança unidade de medida
