+++ trilha-reactjs/package.json
@@ -9,6 +9,7 @@
     "@types/jest": "^26.0.15",
     "@types/node": "^12.0.0",
     "@types/react": "^17.0.0",
+    "bootstrap": "^5.0.2",
     "classnames": "^2.3.1",
     "firebase": "^8.6.8",
     "node-sass": "^5.0.0",
