+++ trilha-reactjs/src/components/Button/styles.scss
@@ -1,10 +1,10 @@
 .button {
-    height: 50px;
-    border-radius: 8px;
+    height: 3.125rem;
+    border-radius: 0.5rem;
     font-weight: 500;
     background: #825afd;
     color: #FFF;
-    padding: 0 32px;
+    padding: 0 2rem;
     
     display: flex;
     justify-content: center;
@@ -16,12 +16,12 @@
     transition: filter 0.2s;
 
     img {
-        margin-right: 8px;
+        margin-right: 0.5rem;
     }
 
     &.outlined {
         background: #FFF;
-        border: 1px solid #835afd;
+        border: 0.063rem solid #835afd;
         color: #835afd;
       }
