--- trilha-reactjs/src/components/RoomCode.tsx
@@ -1,24 +1,0 @@
-import '../styles/roomCode.scss';
-
-import copyImg from '../assets/images/copy.svg';
-
-
-type RoomCodeProps = {
-    code: string;
-}
-
-export function RoomCode(props: RoomCodeProps) {
-
-    function copyRoomCodeToClipboard() {
-        navigator.clipboard.writeText(props.code);
-    }
-
-    return (
-        <button className="room-code" onClick={copyRoomCodeToClipboard}>
-        <div>
-          <img src={copyImg} alt="Copy room code" />
-        </div>
-        <span>Sala #{props.code}</span>
-      </button>
-    )
-}
