+++ trilha-reactjs/src/index.tsx
@@ -2,9 +2,15 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 
+import 'bootstrap/dist/css/bootstrap.min.css';
+
 import './services/firebase';
 
 import './styles/global.scss';
+
+
+
+
 
 ReactDOM.render(
   <React.StrictMode>
