+++ trilha-reactjs/src/pages/Home/styles.scss
@@ -1,0 +1,131 @@
+aside {
+    background: #835afd;
+    color: #fff;
+
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    
+    img {
+        max-width:  20rem;
+    }
+
+    strong {
+        font: 700 2.25rem 'Poppins', sans-serif;
+        line-height: 2.625rem;
+        margin-top: 1rem
+    }
+
+    p {
+        font-size: 1.5rem;
+        line-height: 2rem;
+        margin-top: 1rem;
+        color: #f8f8f8;
+    }
+}
+main {
+    flex: 8;
+    padding: 0 2rem;
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+}
+
+.main-content {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    max-width: 20rem;
+    align-items: stretch;
+    text-align: center;
+
+    > img {
+        align-self: center;
+    }
+
+    h2 {
+        font-size: 1.5rem;
+        margin: 4rem 0 1.5rem;
+        font-family: 'Poppins', sans-serif;
+    }
+
+    form {
+        input {
+            height: 3.125rem;
+            border-radius: 0.5rem;
+            padding: 0 1rem;
+            background: #FFF;
+            border: 1px solid #a8a8b3;
+        }
+
+        button {
+            margin-top: 1rem;
+        }
+
+        button, input {
+            width: 100%;
+        }
+    }
+
+    p {
+       font-size: 0.875rem;
+       color: #737380; 
+
+       a {
+           color: #e559f9;
+       }
+    }
+}
+
+.create-room {
+    margin-top: 4rem;
+    height: 3.125rem;
+    border-radius: 0.5rem;
+    font-weight: 500;
+    background: #ea4335;
+    color: #FFF;
+    
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    cursor: pointer;
+    border: 0;
+
+    transition: filter 0.2s;
+
+    img {
+        margin-right: 0.5rem;
+    }
+
+    &:hover {
+        filter: brightness(0.9);
+    }
+}
+
+.separator {
+    font-size: 0.875rem;
+    color: #a8a8b3;
+
+    margin: 2rem 0;
+    display: flex;
+    align-items: center;
+
+    &::before {
+        content: '';
+        flex: 1;
+        height: 0.063rem;
+        background: #a8a8b3;
+        margin-right: 1rem;
+    }
+
+    &::after {
+        content: '';
+        flex: 1;
+        height: 0.063rem;
+        background: #a8a8b3;
+        margin-left: 1rem;
+    }
+}
