+++ trilha-reactjs/src/pages/Room/styles.scss
@@ -1,0 +1,130 @@
+#page-room {
+    header {
+        padding: 1.5rem;
+        border-bottom: 1px solid #e2e2e2;
+
+        .content {
+            max-width:  70rem;
+            margin: 0 auto;
+            display: flex;
+            justify-content: space-between;
+            align-items: center;
+
+            > img {
+                max-height: 2.813rem;
+            }
+
+            > div {
+                display: flex;
+                gap: 1rem;
+
+                button {
+                    height:  2.5rem;
+                }
+            }
+
+            .powerOff {
+                width: 1.125rem;
+                background: transparent;
+                border: none;
+            }
+        }
+    }
+
+    main  {
+        max-width: 50rem;
+        margin: 0 auto;
+
+        .container {
+            .room-title {
+                margin: 2rem 0 1.5rem;
+                display: flex;
+                align-items: center;
+    
+                h1 {
+                    font-family: "Poppins",sans-serif;
+                    font-size: 1.5rem;
+                    color: #29292e;
+                    font-weight: bold;
+                }
+                
+        
+                span {
+                    margin-left: 1rem;
+                    background: #e559f9;
+                    border-radius: 9999px;
+                    padding: 0.5rem 1rem;
+                    color: #fff;
+                    font-weight: 500;
+                    font-size: 0.875rem;
+                }
+            }
+    
+            form {
+                textarea {
+                    width: 100%;
+                    border: 0;
+                    padding: 1rem;
+                    border-radius: 0.5rem;
+                    background: #fefefe;
+                    box-shadow: 0 0.125rem 0.75rem rgba(0, 0, 0, 0.04);
+                    resize: vertical;
+                    min-height: 8.125rem;
+                }
+            }
+    
+            .form-footer {
+                display: flex;
+                justify-content: space-between;
+                align-items: center;
+                margin-top: 1rem;
+    
+                .user-info {
+                    display: flex;
+                    align-items: center;
+    
+                    img {
+                        width: 2rem;
+                        height: 2rem;
+                        border-radius: 50%;
+                    }
+                    
+                    > span {
+                        margin-left: 0.5rem;
+                        color: #29292e;
+                        font-weight: 500;
+                        font-size: 0.875rem;
+                    }
+                }
+    
+                span {
+                    font-size: 0.875rem;
+                    color: #737380;
+                    font-weight: 500;
+    
+                    button {
+                        background-color: transparent;
+                        border: 0;
+                        color: #835afd;
+                        text-decoration: underline;
+                        font-size: 0.875rem;
+                        font-weight: 500;
+                        cursor: pointer;
+                    }
+                }
+            } 
+    
+            .question-list {
+                margin-top: 2rem;
+            }
+        }
+
+        .alert-info {
+           
+            font-family: 'Poppins', sans-serif;
+            border: none;
+            background-color: rgba(131, 90, 253, 0.3);
+        }
+      
+    }
+}
